Update 11/27/2014 10:18pm
Updated main.py with the beginnings of the section for the mlp based
reinforcement learning (all commented out for now). Also updated game.py
to include a definition for play_with_self_mlp. Also updated QBot to
require a "format" as input for train_update, and train_move so one can
specify either 'TD' or 'mlp'. Started writing the code for the mlp into
these sections.